requirements.txt
yfinance==0.2.28
alpaca-trade-api==3.0.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pandas==2.1.4
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.2
openai==1.6.1
google-generativeai==0.3.2
python-dateutil==2.8.2

